AI PROJECT DATA ANALYST/data_pipeline_agent/requirement.txt
streamlit
langchain
langchain-community
langchain-openai
langgraph
openai
pandas
plotly
python-dotenv
sqlalchemy